refactor: reorder plugins in components module
