+++ packages/nuxt/src/components/plugins/transform.ts
@@ -7,8 +7,8 @@ import { parseURL } from 'ufo'
 import { parseQuery } from 'vue-router'
 import { normalize, resolve } from 'pathe'
 import { genImport } from 'knitwork'
-import { distDir } from '../dirs'
-import type { getComponentsT } from './module'
+import { distDir } from '../../dirs'
+import type { getComponentsT } from '../module'
 
 const COMPONENT_QUERY_RE = /[?&]nuxt_component=/
